+++ Investissement_WebClient.Core/Modeles/DTO/VariationsDto.cs
@@ -2,11 +2,11 @@ namespace Investissement_WebClient.Core.Modeles.DTO;
 
 public class VariationsDto
 {
-    public double VariationPrix24H { get; set; }
+    public decimal VariationPrix24H { get; set; }
     
-    public double VariationPrix7J { get; set; }
+    public decimal VariationPrix7J { get; set; }
     
-    public double VariationPrix1M { get; set; }
+    public decimal VariationPrix1M { get; set; }
     
-    public double VariationPrix1A { get; set; }
+    public decimal VariationPrix1A { get; set; }
 }
